+++ examples/server.rs
@@ -6,7 +6,7 @@ use notify_rust::server::NotificationServer;
 #[cfg(target_os = "windows")]
 fn main() { println!("this is a xdg only feature") }
 
-#[cfg(all(unix, not(target_os = "macos")))]
+#[cfg(all(unix, not(feature = "server"), not(target_os = "macos")))]
 fn main() {
     println!("this is a xdg only feature")
 }
